extracted navigation bar
