+++ jsconfig.json
@@ -14,6 +14,6 @@
     "isolatedModules": true,
     "jsx": "preserve"
   },
-  "include": ["**/**/*.js", "**/**/*.jsx"],
+  "include": ["**/**/*.js", "**/**/**/*.jsx"],
   "exclude": ["node_modules"]
 }
